math: refactor
