+++ src/trx/game/math.h
@@ -2,4 +2,7 @@
 
 #include <trx/game/math/const.h>
 #include <trx/game/math/func.h>
+#include <trx/game/math/geom.h>
+#include <trx/game/math/trig.h>
 #include <trx/game/math/types.h>
+#include <trx/game/math/util.h>
