+++ src/trx/game/math/trig.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdint.h>
+
+int32_t Math_Cos(int32_t angle);
+int32_t Math_Sin(int32_t angle);
+int32_t Math_Atan(int32_t x, int32_t y);
