+++ src/trx/game/math/util.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <trx/game/math/types.h>
+
+bool Bounds32_Intersect(const BOUNDS_32 *a, const BOUNDS_32 *b);
